+++ benchmarks/datascript-bench/bench.clj
@@ -153,7 +153,7 @@
   [
    ;; ["latest" "datomic"]
    ;; ["latest" "datascript"]
-   ;; ["0.9.17" "datalevin"]
+   ["0.9.18" "datalevin"]
    ["latest" "datalevin"]
    ])