+++ src/datalevin/storage.clj
@@ -209,6 +209,7 @@
   (advance-max-gt [this])
   (max-tx [this])
   (advance-max-tx [this])
+  (aid [this attr])
   (max-aid [this])
   (schema [this] "Return the schema map")
   (rschema [this] "Return the reverse schema map")
@@ -235,6 +236,7 @@
   (actual-a-size [this a]
     "Return the number of datoms with the given attribute")
   (start-sampling [this])
+  (analyze [this a])
   (e-sample [this a]
     "Return a sample of eids sampled from full value range of an attribute")
   (v-size [this v]
@@ -244,6 +246,8 @@
   (av-range-size [this a lv hv] [this a lv hv cap]
     "Return the numbers of datoms with given a and v range")
   (cardinality [this a]
+    "Return the number of distinct values of an attribute, an estimate")
+  (actual-cardinality [this a]
     "Return the number of distinct values of an attribute")
   (head [this index low-datom high-datom]
     "Return the first datom within the given range (inclusive)")
@@ -607,14 +611,15 @@
             (mapv first attrs-v)))
 
 (defn- ea->r
-  [schema lmdb e a]
-  (when-let [aid (:db/aid (schema a))]
-    (when-let [^ByteBuffer bf (lmdb/near-list lmdb c/eav e aid :id :int)]
+  [store e a]
+  (when-let [aid (aid store a)]
+    (when-let [^ByteBuffer bf (lmdb/near-list (.-lmdb store) c/eav e aid
+                                              :id :int)]
       (when (= ^int aid ^int (b/read-buffer bf :int))
         (b/read-buffer (.rewind bf) :avg)))))
 
 (declare insert-datom delete-datom fulltext-index check transact-opts
-         ->SamplingWork e-sample*)
+         ->SamplingWork e-sample* analyze*)
 
 (deftype Store [lmdb
                 search-engines
@@ -659,6 +664,8 @@
   (max-tx [_] max-tx)
 
   (advance-max-tx [_] (set! max-tx (inc ^long max-tx)))
+
+  (aid [_ attr] (get (schema attr) :db/aid))
 
   (max-aid [_] max-aid)
 
@@ -712,7 +719,7 @@
     (if (populated?
           this :ave (d/datom c/e0 attr c/v0) (d/datom c/emax attr c/vmax))
       (u/raise "Cannot delete attribute with datoms" {})
-      (let [aid ((schema attr) :db/aid)]
+      (let [aid (.aid this attr)]
         (lmdb/transact-kv
           lmdb [(lmdb/kv-tx :del c/schema attr :attr)
                 (lmdb/kv-tx :put c/meta :last-modified
@@ -791,27 +798,28 @@
   (e-size [_ e] (lmdb/list-count lmdb c/eav e :id))
 
   (a-size [this a]
-    (if-let [aid (get (schema a) :db/aid)]
+    (if-let [aid (.aid this a)]
       (let [^long ms (lmdb/get-value lmdb c/meta aid :int :id)]
-        (or (when (and ms (not (zero? ms)))
-              ;; zero a-size in meta disrupts query, always actual count
-              ms)
+        (or (when (and ms (not (zero? ms))) ms)
             (.actual-a-size this a)))
       0))
 
-  (actual-a-size [_ a]
-    (if-let [aid (get (schema a) :db/aid)]
-      (let [as (lmdb/key-range-list-count
-                 lmdb c/ave
-                 [:closed
-                  (datom->indexable schema (d/datom c/e0 a nil) false)
-                  (datom->indexable schema (d/datom c/emax a nil) true)] :av)]
+  (actual-a-size [this a]
+    (if-let [aid (.aid this a)]
+      (let [c  (.cardinality this a)
+            as (if (< c c/list-dbi-count-divider)
+                 (lmdb/key-range-list-count
+                   lmdb c/ave
+                   [:closed
+                    (datom->indexable schema (d/datom c/e0 a nil) false)
+                    (datom->indexable schema (d/datom c/emax a nil) true)] :av)
+                 (.size this :ave (d/datom c/e0 a c/v0) (d/datom c/emax a c/vmax)))]
         (lmdb/transact-kv lmdb [[:put c/meta aid as :int :id]])
         as)
       0))
 
   (e-sample [this a]
-    (let [aid ((schema a) :db/aid)]
+    (let [aid (.aid this a)]
       (or (when-let [res (not-empty
                            (lmdb/get-range lmdb c/meta
                                            [:closed-open [aid 0]
@@ -833,6 +841,13 @@
                                  TimeUnit/SECONDS)
         (vreset! sample-started true))))
 
+  (analyze [this a]
+    (if a
+      (analyze* this a)
+      (doseq [attr (remove (set (keys c/implicit-schema)) (keys schema))]
+        (analyze* this attr)))
+    :done)
+
   (v-size [_ v] (lmdb/list-count lmdb c/vae v :id))
 
   (av-size [_ a v]
@@ -840,21 +855,36 @@
       lmdb c/ave (datom->indexable schema (d/datom c/e0 a v) false) :av))
 
   (av-range-size ^long [this a lv hv] (.av-range-size this a lv hv nil))
-  (av-range-size ^long [_ a lv hv cap]
-    (lmdb/key-range-list-count
-      lmdb c/ave
-      [:closed
-       (datom->indexable schema (d/datom c/e0 a lv) false)
-       (datom->indexable schema (d/datom c/emax a hv) true)]
-      :av cap))
+  (av-range-size ^long [this a lv hv cap]
+    (if (< (.cardinality this a) c/list-dbi-count-divider)
+      (lmdb/key-range-list-count
+        lmdb c/ave
+        [:closed
+         (datom->indexable schema (d/datom c/e0 a lv) false)
+         (datom->indexable schema (d/datom c/emax a hv) true)]
+        :av cap)
+      (size this :ave (d/datom c/e0 a lv) (d/datom c/emax a hv) cap)))
 
-  (cardinality [_ a]
-    (lmdb/key-range-count
-      lmdb c/ave
-      [:closed
-       (datom->indexable schema (d/datom c/e0 a nil) false)
-       (datom->indexable schema (d/datom c/emax a nil) true)]
-      :av))
+  (cardinality [this a]
+    (if-let [aid (.aid this a)]
+      (let [^long c (lmdb/get-value lmdb c/meta [aid :cardinality]
+                                    [:long :keyword] :id)]
+        (or (when (and c (not (zero? c))) c)
+            (.actual-cardinality this a)))
+      0))
+
+  (actual-cardinality [this a]
+    (if-let [aid (.aid this a)]
+      (let [c (lmdb/key-range-count
+                lmdb c/ave
+                [:closed
+                 (datom->indexable schema (d/datom c/e0 a nil) false)
+                 (datom->indexable schema (d/datom c/emax a nil) true)]
+                :av)]
+        (lmdb/transact-kv lmdb [[:put c/meta [aid :cardinality] c
+                                 [:long :keyword] :id]])
+        c)
+      0))
 
   (head [this index low-datom high-datom]
     (retrieved->datom lmdb attrs
@@ -939,12 +969,12 @@
   (av-first-datom [this a v]
     (when-let [e (.av-first-e this a v)] (d/datom e a v)))
 
-  (ea-first-datom [_ e a]
-    (when-let [r (ea->r schema lmdb e a)]
+  (ea-first-datom [store e a]
+    (when-let [r (ea->r store e a)]
       (avg-retrieved->datom lmdb attrs e r)))
 
-  (ea-first-v [_ e a]
-    (when-let [r (ea->r schema lmdb e a)] (avg-retrieved->v lmdb r)))
+  (ea-first-v [store e a]
+    (when-let [r (ea->r store e a)] (avg-retrieved->v lmdb r)))
 
   (v-datoms [_ v]
     (mapv #(ae-retrieved->datom attrs v %)
@@ -1051,9 +1081,9 @@
       out))
 
   (eav-scan-v
-    [_ in out eid-idx attrs-v]
+    [store in out eid-idx attrs-v]
     (when (seq attrs-v)
-      (let [attr->aid #((schema %) :db/aid)
+      (let [attr->aid #(aid store %)
             get-aid   (comp attr->aid first)
             attrs-v   (sort-by get-aid attrs-v)
             aids      (mapv get-aid attrs-v)
@@ -1091,9 +1121,9 @@
            (b/indexable nil (aget aids (dec (alength aids)))
                         c/vmax nil c/gmax)] :avg))))
 
-  (eav-scan-v-list [_ in eid-idx attrs-v]
+  (eav-scan-v-list [store in eid-idx attrs-v]
     (when (seq attrs-v)
-      (let [attr->aid #((schema %) :db/aid)
+      (let [attr->aid #(aid store %)
             get-aid   (comp attr->aid first)
             attrs-v   (sort-by get-aid attrs-v)
             aids      (mapv get-aid attrs-v)
@@ -1272,6 +1302,12 @@
                              ts))
     ts))
 
+(defn- analyze*
+  [store attr]
+  (when-let [aid (aid store attr)]
+    (actual-cardinality store attr)
+    (e-sample* store attr aid (actual-a-size store attr))))
+
 (defn sampling
   "sample a random changed attribute at a time"
   [^Store store]
@@ -1283,7 +1319,7 @@
         (when (and (not (closed? store))
                    (< (* ^long (a-size store attr) ^double c/sample-change-ratio)
                       acount))
-          (e-sample* store attr aid (actual-a-size store attr))
+          (analyze* store attr)
           (.put counts aid 0))))))
 
 (deftype SamplingWork [^Store store]